relocate dynamic APIs
These files are about to change. This commit just moves them to make diff and history tracking easier to follow
