+++ packages/next/headers.d.ts
@@ -1,1 +1,1 @@
-export * from './dist/client/components/headers'
+export * from './dist/server/request/headers'
