+++ packages/next/src/api/headers.ts
@@ -1,1 +1,1 @@
-export * from '../client/components/headers'
+export * from '../server/request/headers'
