+++ packages/next/src/client/components/router-reducer/compute-changed-path.ts
@@ -3,7 +3,7 @@ import type {
   Segment,
 } from '../../../server/app-render/types'
 import { INTERCEPTION_ROUTE_MARKERS } from '../../../server/lib/interception-routes'
-import type { Params } from '../params'
+import type { Params } from '../../../server/request/params'
 import {
   isGroupSegment,
   DEFAULT_SEGMENT_KEY,
