+++ packages/next/src/client/components/static-generation-async-storage.external.ts
@@ -3,7 +3,7 @@ import type { IncrementalCache } from '../../server/lib/incremental-cache'
 import type { DynamicServerError } from './hooks-server-context'
 import type { FetchMetrics } from '../../server/base-http'
 import type { Revalidate } from '../../server/lib/revalidate'
-import type { FallbackRouteParams } from './fallback-params'
+import type { FallbackRouteParams } from '../../server/request/fallback-params'
 
 // Share the instance module in the next-shared layer
 import { staticGenerationAsyncStorage } from './static-generation-async-storage-instance' with { 'turbopack-transition': 'next-shared' }
