+++ packages/next/src/server/app-render/entry-base.ts
@@ -19,21 +19,17 @@ import { ClientPageRoot } from '../../client/components/client-page'
 import {
   createUntrackedSearchParams,
   createDynamicallyTrackedSearchParams,
-} from '../../client/components/search-params'
-import { createDynamicallyTrackedParams } from '../../client/components/fallback-params'
+} from '../request/search-params'
+import { createDynamicallyTrackedParams } from '../request/fallback-params'
 import * as serverHooks from '../../client/components/hooks-server-context'
 import { NotFoundBoundary } from '../../client/components/not-found-boundary'
 import { patchFetch as _patchFetch } from '../lib/patch-fetch'
 // not being used but needs to be included in the client manifest for /_not-found
 import '../../client/components/error-boundary'
 
-import {
-  preloadStyle,
-  preloadFont,
-  preconnect,
-} from '../../server/app-render/rsc/preloads'
-import { Postpone } from '../../server/app-render/rsc/postpone'
-import { taintObjectReference } from '../../server/app-render/rsc/taint'
+import { preloadStyle, preloadFont, preconnect } from './rsc/preloads'
+import { Postpone } from './rsc/postpone'
+import { taintObjectReference } from './rsc/taint'
 
 // patchFetch makes use of APIs such as `React.unstable_postpone` which are only available
 // in the experimental channel of React, so export it from here so that it comes from the bundled runtime
