+++ packages/next/src/server/web-server.ts
@@ -1,7 +1,7 @@
 import type { WebNextRequest, WebNextResponse } from './base-http/web'
 import type RenderResult from './render-result'
 import type { NextParsedUrlQuery, NextUrlWithParsedQuery } from './request-meta'
-import type { Params } from '../client/components/params'
+import type { Params } from './request/params'
 import type { LoadComponentsReturnType } from './load-components'
 import type {
   LoadedRenderOpts,
