+++ test/e2e/app-dir/actions/app/redirect-target/page.js
@@ -1,4 +1,4 @@
-import { cookies } from 'next/dist/client/components/headers'
+import { cookies } from 'next/dist/server/request/headers'
 
 export default function Page() {
   const redirectCookie = cookies().get('redirect')
